Fix a memory leak in the Graphical Layout Editor.
The rendering requires a looper, but never actually uses it.
Some views (like scrollView) make use of messages during rendering
putting messages into the looper message queue which was never
read or emptied.

In the case of the scrollview, the message actually contains
a reference to the scrollview, which would then leak (with all
its children views)

The fix is to delete the looper that was created. This must be
done by reflection as there's no public access to it, and this
must be done from ADT so that all versions of layoutlib get
the fix.

Change-Id: I998ad0ec17e77e36a42d77f8ab888917a9ff6441