Force recompilation after library changes and fix some clean-up/refresh issue in the builders
One of the fix is that the gen folder is no longer deleted on clean (or full build)
which should reduce errors where eclipse complain the gen folder is missing
even though it's present.

Also move ProjectState into SDK as it's really only managed
by the SDK and it allows us to move its constructor to package private.

Change-Id: I321ac23f756652e7082297d2ef560a654e97cf67