Improve library link/unlink again.
Improved library list added to the bundle by comparing
the old and new library list when something changed.
This is the only way to ensure that all indirect libraries
are added removed.

Since it means we now have a list of added and a list
of removed libraries, the link and unlink methods are
merged into a single updateLink method that does it all.

This also means that there's only one type of action
bundle, so the class/fields name have been updated to
reflect the new way it's done.

Change-Id: I808a8f1536171ebebca3a9bb5370e9dc0b6433d5