Add new exceptions to ddmlib.
- AdbCommandRejectedException is thrown when adb doesn't respond
to the command with OKAY. This used to throw a normal IOException
but it can be useful to throw a different type. The message
of the exception is the error string returned by adb.

- ShellCommandUnresponsiveException is the new "timeout" exception
for output received by shell command running on devices. This
makes the distinction between timeout when talking to adb and
issue with shell command not outputting anything. Also made the
javadoc for the IDevice.executeShellCommand clearer to what the
"timeout" (renamed to maxTimeToOutputResponse) does.

Also added a better timeout to the IDevice methods to install/uninstall
apps as the default 5sec timeout was likely to be too low.
Current default value is 2min.

Change-Id: I4ecb9498926295a4e801e71b33df5d611e8120b8