ADT GLE2: Let groovy rules access a public API of the RulesEngine.
This is done by injecting a property in the groovy rule instance
with a new interface, the IClientRulesEngine. The client callback
currently has 2 useful methods: debugPrintf (which has been moved
out of INode) and loadRule(). This last one is the key here, it
allows one rule to request another one to be loaded and get its
object.

Change-Id: I2881854e33cd3b41565dd1e16aaba1484ef765db