SdkManager: new schema with <platform-tool> package.
This adds the new schema (v3), the internal definitions and
constants. However the schema v3 is not enabled as default and
the whole thing is not used by the downloader/installer yet.

Note: some files in this folder are using Windows \r\n end-of-lines.
I will convert them to Unix \n EOLs in *another* CL, not this one.

Change-Id: I327f12571ad3959e85a7b972cad98663c09cce33