Allow adt-tests to access package-private elements.
Previously when running tests in the continuous tests system, adt-tests could
not access any package-private elements in adt. This was because adt-tests was
a separate plugin and thus had a separate class loader.

With this commit, adt-tests is now a fragment, and its code runs inside the
adt plugin.

Bug 2179353

Change-Id: Iec8c08bed2fa116b38490a774cd3bd6f4c7358a5