SDK Manager: Remove static package class map.
When using the manager in --no-ui mode, one can use
--filter to select a list of package types to install.
This list is based on the array in SdkRepository constants
and it used need a static map to the actual classes to do
the filtering. The static map is now computed at runtime
so that we don't risk forgetting to update it.

Change-Id: I7b091c36942d3b9358595f2194f2edc76c517245