Don't process events for non Android projects
Project opened event are triggered for all events.
The code would then try to get the ProjectState for
the opened project whether it was an Android project
or not, filling the log with errors.

Same thing for file modification event for default.properties.
Only care if the containing project is an Android project.

http://b.android.com/10651

Change-Id: I35d7b546794ced41cffd382a9693294cfeb3e311