DDMS now receives the adb location through an extension point.
Previously DDMS received the adb location through a normal Java API
which requires the other plug-in to be actually running (This
was done in the the start method of ADT).

The new change allows DDMS to start a plug-in (by loading
an extension provided by the plug-in) to query for the location
of adb.

This allows us to have plug-ins with no UI able to provide
the location of ADB (ie a "platform" plug-in that has no UI
besides a pref page, that is started by DDMS indirectly).

Also cleaned up how HierarchyViewer use the AndroidDebugBridge.

Change-Id: I8e842a294eea94c06417149144a2ce435e719cfd