New zoom controls in the layout editor.
New 100% and "real size" buttons. The real size
computes a scale value based on the dpi of the monitor
and the dpi of the layout device info.

Also cleaned-up the CustomToggle class to be a
CustomButton (toggle optional) as the 100% and
zoom in/out buttons are not toggles.

Migrated the clipping button to be set up through
a CustomButton.

Added groups of CustomButton, which are closer to
each others than normal.

Change-Id: Id4b6ed54b7f275f848333b04aeb42ef80e89ea9b