GLE2: Properly represent mixed choices in dynamic context menu.
Example: when display the context menu for a multiple selection,
some items might be on wrap_parent whilst others might be in
match_parent. This new change has the menu display a different
icon for these choices and indicate that N out of M items are
in a given state. Basically this converts boolean support to
a tristate support.

Change-Id: Ia409a87fa17e5d36912ace6dde5f4eb094fa865f