GLE2: fix layout editor not properly closed when dirty.
The symptom was that once the layout editor had a modification
and was in the "dirty" mode, closing it would keep the model
around and re-opening would show the previous state. That was
a red herring.

The actual issue is that the undo begin/end wrapper methods
were calling getModelForEdit() without doing a proper
releaseFromEdit() call after the fact. Somewhere down in the
overly complex code that disposes the EditorPart there is
a test that would not purge the model if it is still locked.

The fix is thus in the begin/endUndoRecording of the base
AndroidXmlEditor.

This CL also cleans up some IFile usage; when fixing the
code above I noticed we can now get the model without
first getting an SSE internal document, instead we can use
the editor's IFile. However I doubt SSE for 3.3 would have
the API, I need to check that in another CL. Later, it's a P3.

Change-Id: I2437475dfeee9d6689b7b604782ae140d7aff1c3