Use proguard for release builds through Ant.
This is only activated if default.properties contains
a property "proguard.config" with the name of a proguard
config file.

Some clean-up in the Ant tasks and in the name
of the properties used by the rules and the custom
tasks to make them clearer.

Added a new test app with a project using a jar file
as well as a library using a jar file.

Change-Id: Ia8f4d873025993d454c0a484e61d47ae679ea79c