ADT: Fix attributes UI in Android Manifest Editor.
External bug: http://b.android.com/11392

Issue: all customized fields were showing as regular "string"
references (with the default string browser) instead of having
customized pre-selected choices, etc.

The core issue is that the UI widgets were created using reflection
but I changed the base class constructor of the default
TextAttributeDescriptor recently and forgot to update DescriptorUtils
accordingly. Since it was failing to create a custom widget it was
reverting to the default UI for a string reference.

The long-term fix is to stop using reflection.
Creator pattern to the rescue!
Instead there's a new ITextAttributeCreator interface and the
override map uses a static "creator" instance that will instantiate
the correct UI widget when needed.

Change-Id: Ie7afbf682a2aa8ae18565d08445e73a8c051a6c7