Add drawing styles to the layout helper API
Pull color constants out of the specific layout helper classes (groovy
scripts) and the canvas editor and use a generic style enum instead in
the interface, and associate the visual attributes (color, line style,
thickness, alpha) with a swt-specific enum on the editor side. 	There
is a single new	API method which takes an enum parameter, which	should
let us add drawing styles over time. By	having the color definitions
on the tool side rather	in the specific	layout helpers it's not	only
easier to change the colors but	also easier to ensure that the
different helper all stay consistent as we change color schemes.

In the immediate term (next integration) I'll change some of the
colors; after that we should make the colors adapt to the chosen
theme, and eventually these should be provided via the SDK from the
themes themselves.

This changeset doesn't actually change the colors used for the various
types of visual feedback (selection, hover, drop-zone, etc) - I'll
investigate that next. For that reason I also didn't replace all the
various client-side color usage in the RelativeLayout.

Change-Id: Iddf4ace9006ec02d9907c3c37d539ab7414f1371