SDK Manager: manage URL sites grouped in categories.
Sources are now grouped in categories:
- android official site
- all 3rd party addons which list is remotely fetched
- local overrides (from getenv SDK_UPDATER_(USER_)URLS)
- user-added add-on URLs

SdkSources (plural Sources) is the collection of all URL
sites (each one being an SdkSource that contains a collection
of packages).
SdkSourceCategory is the category of a given SdkSource.

Each source has its download URL but also an UI-visible name
which is displayed if known -- this is used for our default
android source and for 3rd party remote addons.
For all source overrides (e.g. user added or from the getenv
vars) we'll simply display the URL in the tree.

Change-Id: Idc3e9ba3395818e33fc7505ead7b0bd0be867ea3