SDK Manager: simplify getenv var to override URLs.
SDK_TEST_URLS is a semicolon-separated list of URLs that can be used to
seed the SDK Updater list. This is only meant as a debugging and QA testing
tool and not for user usage.
To be used, the URLs must either end with the / or end with the canonical
filename expected for an addon list, addon repo or full repo.
This lets QA use URLs ending with / to cover all cases.

Change-Id: Ia7232d5d8a5b6e85d98e735b93a969e42624e919