Improve visual feedback for relative layout and frame layout.
Improve the visual feedback for the relative layout and the frame
layouts:

1. When	there is multiple selection, don't draw	adjacent boxes (since
that's not how things will look once you drop); instead, draw
overlapped boxes, For example if you drag a button and a clock, the
button will be placed on top of the clock, not next to it, in both
layouts.

2. Replace the center oval+x pattern with the stippled orange line
pattern, and make some fixes to relative layout such that (when we
have bounds) the preview rectangle is positioned correctly for the
parent-relative bounds.

Change-Id: Ie4c7b127cb28d9fedf194e9067f2fa9c82ede44e