Fix drag and drop handler to handle reentry
If you drag from the palette over a linear layout view, you get drop
feedback. However, if you drag outside the linear layout and then back
in, you often don't get any more feedback. This happens pretty easily
with nested layouts.

The problem is that the feedback object is initialized and reset in
onDropEnter, but the actual guideline computation happens in
onDropMove (since only drop move is handed the mouse coordinates). The
bug happens because when you leave and return to the drop zone, the
feedback object is reset, and then the drop feedback is painted - all
before the onDropMove code has been called again.

The fix is simple: On drop enter, also call onDropMove immediately
such that the layout helpers are given a chance to initialize
themselves with the entry mouse position, before they are consulted
for drop feedback data.

This changeset also contains a fix for LinearLayouts: When there are
no children, place the guideline at x=0 or y=0 (depending on whether
the layout is horizontal or vertical). Without this fix, the line is
placed in the center of the view, which is misleading since that is
not where the view will be placed upon drop.

Change-Id: Ib5e17a2d9d3818677e4209126233bbde527207c8