SDK Manager: fix to suggest platform-tools install.
The way the updater currently works, it will only
generate the source.props of a new package based on the
attributes it knows from that package. That means mean
Tools r7 is updating tools, it will not add the proper
min-platform-tools-rev to the new Tools r8 package.

When "Update All" is selected, we try to do 2 new things:
- make sure to lool at local *existing* packages for
  potential missing dependencies, and suggest them for
  install.
- if a package doesn't have a min-platform-tools-rev set,
  suggest the higest revision available.

Change-Id: I76cdbc2818133429b2726d3127eedd7e65579a7e