Prevent accidental dragging of the root layout
The layout editor deliberately does not reset the selection if you
start dragging within the bounding rectangle of one (of possibly
multiple) selected widgets. However, since it's pretty easy to have
the root element selected, this sometimes means you end up trying to
drag an item only to realize you can't because everything fits within
the root bounding rectangle.

This changeset fixes this usability issue by making the root widget a
special case. If you are dragging *only* the root, things work as
before. But if you are trying to drag anything else, it will ignore
the root, which means that you can drag one or more items
independently of the root whether or not it is included in the
selection.

Change-Id: Ia7497659e57dcc6431c3760a53847268a8b90815