Hide anchor lines and tweak text positioning
Temporarily hide the anchor lines (the lines from each selected
element to its parent's bounds edges) until we can find a way to make
it clearer what these mean.

Also tweak the pixel positioning of the selection label text to
account for the padding around the textbox mask since it was partially
overlapping the selection stroke.

Change-Id: Iaa4c7139d4d56288b78a104647d5cbb5752f7888