SDK Manager: Fix dependencies when installing selected packages.
In the SDK Manager, when you select a specific package from an
available repository to install it, it used to not notice if there
was a broken depedency. This fixes it. It means we have to force
loading all known available repo sources to find them if necessary.

This CL fixes a minor issue: trim all URLs before trying to use
them. The CL is voluntarily a bit defensive for that. This can
happen when we load URLs from external addon sites which we do
not control.

Change-Id: I652e1fa0e74b4384d0e95819970a59519fafa05a