Fix ddms filechooser behavior and Mac accelerator
Fix a couple of issues in ddms:

First, fix the code to remember the path you last opened the file
chooser with; if there is nothing in $lastImageSaveDir, it should look
at $imageSaveDir (which is the variable the default setting
($user.home) is initialized into).

Second, and this seems to be Mac specific, the
FileDialog.getFilterPath() call does not return the path you have
navigated to, which means that on the Mac it never sets
$lastImageSaveDir correctly - it always sets it to the original
suggestion. The fix is trivial - use File#getParent instead which does
the String manipulation to extract the parent portion of a string
which represents a path.

Finally, on Macs (only), make the keybindings use the Command key
instead of the Control key since that's the norm.

Change-Id: I1b0f381606f5373ddad973754e49ce07856a9bae