Update application icons for sdkmanager and ddms
Update the dock icon used on Macs for the sdkmanager to one that has
transparency, just as the emulator icon already does. This is the norm
on Macs.

Update the application icon used for ddms to be	an android with a
wrench, in a similar style to the other tool icons.

This icon is also transparent, which I hope will work on Windows - it
would be great if Raphael could verify that (and I should probably try
to get a Windows image into my VMWare install so I can do these checks
myself in the future.)

Change-Id: Icd0954cb77ba0845034a79bf9df6dd271019632e