Add gesture support, marquee selection, and refactoring
This checkin adds support for gestures and overlays. Gestures are
sessions of mouse/keyboard activity, and this is documented in the
javadoc for the new Gesture class. Overlays are units of graphics, and
these are documented in the Overlay javadoc. The gesture architecture
lets us isolate the logic for each different type of operation
(marquee, resize, move, etC), and with associated overlays we don't
attempt to for example paint drag feedback during a resize operation,
etc.

The checkin also adds marquee selection (as a second gesture, in
addition to the existing drag & drop based move gesture), along with
some associated changes in how the root view is treated.

As part of isolating the mouse handling and painting related to
gestures, painting etc., I also refactored the code quite a bit.
LayoutCanvas which used to be a large class has been split into a
number of new classes, one for each area of responsibility:

- The mouse listener and drag & drop code has been moved into a
  GestureManager. (A lot of the	drop handling code also	came from the
  CanvasDropListener class.)

- Code related to maintaining the set of rendered views, and
  performing searches in the views, has been moved into a
  ViewHierarchy class.

- Code related to selection has been moved into a SelectionManager.

- Various individual painting pieces (outline, hover, etc) have been
  moved into individual Overlay classes such as OutlineOverlay,
  HoverOverlay, SelectionOverlay, etc. This also moved associated
  resource allocation and cleanup into the overlays.

- New coordinate classes, ControlPoint and LayoutPoint, are used
  instead of ints and plain Points to make it really clear which
  methods require coordinates in the layout (such as the
  ViewHieararchy search methods) and which ones require coordinates in
  the canvas control (such as paint methods). There are factory methods
  to automatically construct the right kind of coordinate from
  different types of mouse events, as well as methods to convert
  between the two.

I also tweaked the visual appearance of selection a bit more, and
some other misc cleanup.

Change-Id: I666aabdcd36720bebe406b68237e8966d985fb8f