Fix icon painting transparency for the palette
The icons shown in the palette do not have transparency applied to
them, which means that you end up with white squares on a white
background. This looks okay in the outline view where the icons are
shown on a white background, but it does not look good in the palette
where they are on a gray background.

This changeset fixes the icon generation a bit such that it records a
transparent pixel for these icons. It's not an ideal solution since we
only have on/off transparency, not an alpha channel, which means that
we can't have a smoothly anti-aliased edge of the circles regardless
of the background. Luckily, a light color seems to work fine for both
cases.

While I was at it I also adjusted the font positioning slightly on the
Mac platform for some letters that needed it.

Change-Id: Idf5a4f6709096935122bfe05cfecffc63e74a863