Improvements to the new Layoutlib API.
Move to classes instead of interfaces so that the layoutlib
API code can provide default implementation.

This will reduce the testing on the ADT side about whether a method
is present for the current API level or not.

Also moved all the parameters of startLayout (actually renamed createScene)
into a SceneParams class.

Change-Id: I58389cd1bed9e79f6825b7c4e5a79206496ca439