Change the adbLocation ddms extension to provide more tools location.
Previously DDMS used the adb location to find the location of other tools,
but adb moved to a different tool folder.

adbLocator extension changed to toolsLocation and provide explicit
locations for all the tools DDMS cares about. This way the logic
of the tools location is only in ADT instead of being duplicated
in DDMS.

Change-Id: I87f19c7705cb822dc793264f11e680ba09eb7f40