Win SDK: Split the win-sdk makefile and script into 2 parts. DO NOT MERGE.
[Merge Change Ic94d261b from master_open into tools_r8 to fix broken SDK build.]

The main makefile+scripts to build the Windows SDK is still
in development/build/tools. However it defers to a new
matching set of files here in sdk/build to build and
package things that depend on the sdk.git or external/qemu.git.

This will make it easier for us to prepare SDKs based on
a tools_rN branch that isn't cut at the same time than the
platform branch.

This is a multi-part changeset. The other part is in
development.git.

Change-Id: I8072bf59ee1af57826a8bafa0103afd6fe3fc858