LinearLayout guide fixes
When we	have bounds for	the dragged items, we show a preview rectangle
where the item will appear after the drop. We show this rectangle
midway between the two siblings that share the insert
position. However, when you are inserting *after* the last item in the
LinearLayout, there is no reason to show the item midway, since at
this point nothing below the insert position needs to be shifted
down. This checkin changes this such that for the last insert
position, we show the rectangle fully below the insert line (or to the
right of it in case of a horizontal layout).

This changeset also fixes a bug	where the last (available, not active)
dropzone would not be shown on a palette drag.

Change-Id: If449b43582e072c9e0ad6d7741afbe8e845e8df9